feat(servo): use float instead of uint16 for servo position
Allows higher accuracy and is consistent with the steering value.

+++ apps/servo/include/servo.h
@@ -18,8 +18,8 @@ typedef struct {
   uint16_t current;
   float steering_angle;
   uint16_t steering_pulse;
-  int16_t position;  // Reported position
-  bool reverse;      // Reverse steering direction
+  float position;  // Reported position
+  bool reverse;    // Reverse steering direction
 } servo_state_t;
 
 typedef enum {
